+++ examples/Example/Time/Compat.hs
@@ -29,7 +29,7 @@ transpose oldTime =
   T.ZonedTime (T.LocalTime day tod) (T.TimeZone tzmin _ _) = oldTime
 
   newDate :: H.Date
-  newDate = C.dateFromTAIEpoch $ T.toModifiedJulianDay day
+  newDate = C.dateFromMJDEpoch $ T.toModifiedJulianDay day
 
   timeofday :: H.TimeOfDay
   timeofday = C.diffTimeToTimeOfDay $ toRational $ T.timeOfDayToTime tod